chore: move builtins_pod tests to a sibling file
Match the project convention used by pod.rs / pod_tests.rs and
symbols.rs / symbols_tests.rs.
